+++ examples/httprelay/uvlib/mybuf.h
@@ -13,6 +13,10 @@ extern "C" {
 #define MYBUF_DEFAULT_SIZE 4096
 #endif
 
+#if MYBUF_DEFAULT_SIZE < 8
+#error "MYBUF_DEFAULT_SIZE must >= 8"
+#endif
+
 typedef struct {
     char* buf;
     char mybuf[MYBUF_DEFAULT_SIZE];
@@ -23,6 +27,16 @@ void mybuf_init(mybuf_t* buf);
 size_t mybuf_space(mybuf_t* buf);
 void mybuf_reserve(mybuf_t* buf, size_t size);
 void mybuf_append(mybuf_t* buf, const char* data, size_t len);
+
+/*
+ * append hex dump to mybuf
+ * show_header: show 00~0F address header, default is 0
+ * show_address: show 0000~FFFF address on left
+ * show_ascii: show ascii on right if can print, default is 0
+ * def_ascii: when show_ascii but can't print, show def_ascii instead, default is '?'
+ */
+void mybuf_append_hexdump(mybuf_t* buf, const char* data, size_t len,
+                          int show_header, int show_address, int show_ascii, char def_ascii);
 
 #ifdef __GNUC__
 void mybuf_cat_printf(mybuf_t* buf, const char* fmt, ...)