+++ examples/httprelay/uvlib/uv_log.c
@@ -6,8 +6,35 @@
 #include <time.h>
 #include <uv.h>
 
+#include "mybuf.h"
+
 static uv_log_level_t g_level = uv_log_level_info;
 static char g_logfile[1024] = {0};
+
+uv_log_level_t uv_log_level_from_string(const char* str) {
+#define XX(name, value)            \
+    if (strcmp(str, #name) == 0) { \
+        return value;              \
+    }
+    UV_LOG_LEVELS_MAP(XX)
+#undef XX
+
+    return uv_log_level_invalid;
+}
+
+const char* uv_log_level_to_string(uv_log_level_t level) {
+    switch (level) {
+#define XX(name, value) \
+    case value: {       \
+        return #name;   \
+    }
+        UV_LOG_LEVELS_MAP(XX)
+#undef XX
+        default:
+            return "invalid level";
+            break;
+    }
+}
 
 void uv_log_set_level(uv_log_level_t level) {
     g_level = level;
@@ -29,7 +56,7 @@ void uv_log_raw(uv_log_level_t level, const char* msg) {
     size_t off;
     uv_timeval64_t now;
     char buf[64];
-    const char* c = "ADIWEF";
+    const char* c = "AVDIWEF";
     int log_to_stdout = g_logfile[0] == '\0';
     FILE* fp = NULL;
 
@@ -43,6 +70,17 @@ void uv_log_raw(uv_log_level_t level, const char* msg) {
     fprintf(fp, "%s %c %s\n", buf, c[level], msg);
     fflush(fp);
     if (!log_to_stdout) fclose(fp);
+}
+
+void uv_log_hexdump(uv_log_level_t level, const char* buf, size_t len,
+                    int show_header, int show_address, int show_ascii, char def_ascii) {
+    mybuf_t mybuf;
+    if (level < g_level || level > uv_log_level_fatal) return;
+    mybuf_init(&mybuf);
+    mybuf_cat_printf(&mybuf, "hex dump of %p, len=%zu\n", buf, len);
+    mybuf_append_hexdump(&mybuf, buf, len, show_header, show_address, show_ascii, def_ascii);
+    uv_log_raw(level, mybuf.buf);
+    mybuf_clear(&mybuf);
 }
 
 static void uv_log_vprintf(uv_log_level_t level, const char* fmt, va_list ap) {