+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Ademco Change Log
 
+
+## 1.5.0
+
+- `HbComResp_A2_Iter.com` changed to pointer
+- update parser for `timestamp`
+- add `AdemcoControlSource` defs
+
+
+## 1.4.0
+
+- `hb_com reqeust, response` update
+
+
 ## 1.3.1
 
 - compile with warning level 4
 - change `static inline` function to micro
+
 
 ## 1.3.0
 
@@ -11,11 +25,13 @@
 - fix `struct tm not initilized` on parse
 - fix parsing pointer on parse timestamp failed
 
+
 ## 1.2.0
 
 - add config
 - add `offset` for parse error
 - refactor parser
+
 
 ## 1.1.0
 
@@ -23,6 +39,7 @@
 - add ADEMCO_PACKET_ACCT_MAX_LEN
 - hb_com 读写主机账号命令
 
+
 ## 1.0.0
 
-initial version
+- initial version